+++ mayan/apps/documents/utils.py
@@ -1,5 +1,21 @@
 from __future__ import unicode_literals
 
+import pycountry
+
+from django.utils.translation import ugettext_lazy as _
+
+from .settings import setting_language_codes
+
+
+def get_language_choices():
+    return sorted(
+        [
+            (
+            iso639_3, _(pycountry.languages.get(alpha_3=iso639_3).name)
+            ) for iso639_3 in setting_language_codes.value
+        ], key=lambda x: x[1]
+    )
+
 
 def parse_range(astr):
     # http://stackoverflow.com/questions/4248399/